Renamed AuthControllerTests class in XUnit project
